repositories and register and login dtos added

+++ src/main/java/org/example/data/models/Account.java
@@ -5,12 +5,14 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @Document
 public class Account {
     @Id
     private String id;
     private String accountNumber;
-    private double balance;
+    private BigDecimal balance;
     private User user;
 }
